Move version-specific unit tests into EZSP test files
